Renamed Ndata to Edm
