+++ src/containers/catalog/CatalogEpic.js
@@ -7,7 +7,7 @@ import { SearchApi } from 'loom-data';
 
 import * as actionTypes from './CatalogActionTypes';
 import * as actionFactories from './CatalogActionFactories';
-import * as ndataActionFactories from '../ndata/NdataActionFactories';
+import * as ndataActionFactories from '../edm/NdataActionFactories';
 import { Permission } from '../../core/permissions/Permission';
 import type { EntitySet } from '../../components/entityset/EntitySetStorage';
 import { EntitySetNschema } from '../../components/entityset/EntitySetStorage';
