+++ src/core/redux/ReduxEpic.js
@@ -1,6 +1,6 @@
 import { combineEpics } from 'redux-observable';
 import CatalogEpic from '../../containers/catalog/CatalogEpic';
-import NdataEpic from '../../containers/ndata/NdataEpic';
+import NdataEpic from '../../containers/edm/NdataEpic';
 
 export default function reduxEpic() {
   return combineEpics(
