+++ src/views/Main/routes.js
@@ -14,7 +14,7 @@ import PageConsts from '../../utils/Consts/PageConsts';
 import EnvConsts from '../../utils/Consts/EnvConsts';
 import UserRoleConsts from '../../utils/Consts/UserRoleConsts';
 import StringConsts from '../../utils/Consts/StringConsts';
-import { configure as edmApiConfigure } from '../../containers/ndata/EdmApi';
+import { configure as edmApiConfigure } from '../../containers/edm/EdmApi';
 
 import OrganizationsComponent from '../../containers/organizations/components/OrganizationsComponent';
 import OrganizationDetailsComponent from '../../containers/organizations/components/OrganizationDetailsComponent';
